requirements.txt
uvicorn==0.18.3
pydantic==1.10.2
pymongo==4.2.0
fastapi==0.85.0
python-dotenv==0.21.0
anyio==3.6.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.3.0
h11==0.14.0
idna==3.4
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
passlib==1.7.4
cryptography==43.0.0
openpyxl==3.1.5
python-jose==3.3.0 
bcrypt==4.2.0
python-multipart==0.0.9
pytest==8.3.2